+++ test/unit/common/websocket_test.rb
@@ -1,0 +1,11 @@
+require 'test_helper'
+require 'webmock/minitest'
+
+class AppiumLibCoreTest
+  class WebSocketTest < Minitest::Test
+    def test_connect_websocket
+      ws = ::Appium::Core::WebSocket.new(url: 'ws://localhost:9292')
+      assert_equal nil, ws.client
+    end
+  end
+end